services/source-code-service/requirements.txt
# Core dependencies
fastapi==0.95.1
uvicorn[standard]==0.22.0
pydantic==1.10.7
python-dotenv==1.0.0

# GitHub integration
GitPython==3.1.31
PyGithub==1.58.1

# Database (for future MongoDB integration)
pymongo==4.3.3
motor==3.1.2

# Utilities
requests==2.29.0
aiohttp==3.8.4
python-multipart==0.0.6
Jinja2==3.1.2
email-validator==2.0.0.post2

# Testing
pytest==7.3.1
pytest-asyncio==0.21.0
httpx==0.24.0

# Code analysis (for future)
pygments==2.15.1

# OpenAI integration
openai==0.27.6
tiktoken==0.4.0
